feat:add Empty check and test Company API into result

+++ src/main/java/com/example/employee/restfulapi/controller/CompanyController.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -40,17 +41,28 @@ public class CompanyController {
     }
 
     @RequestMapping(method = RequestMethod.POST)
-    public Company saveCompany(Company company) {
+    public Company saveCompany(Company company) throws Exception {
+        if (company.getCompanyName() == null || company.getEmployeesNumber() == null) {
+            throw new Exception("Invalid Company!");
+        }
         return companyRepository.save(company);
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.PUT)
-    public Company updateCompany(@PathVariable Long id, @ModelAttribute Company company) {
-        return companyRepository.undateById(id, company.getCompanyName(), company.getEmployeesNumber());
+    public Company updateCompany(@PathVariable Long id, @ModelAttribute Company company) throws Exception {
+        if (getCompany(id) == null) {
+            throw new Exception("Company not found by id: " + id);
+        }
+       companyRepository.updateById(id, company.getCompanyName(), company.getEmployeesNumber());
+       return companyRepository.findById(id);
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
-    public Company deleteCompany(@PathVariable Long id) {
-        return companyRepository.deleteById(id);
+    public Company deleteCompany(@PathVariable Long id) throws Exception {
+        if (getCompany(id) == null) {
+            throw new Exception("Company not found by id: " + id);
+        }
+        companyRepository.deleteById(id);
+        return null;
     }
 }
